Lift the artist-watchlist endpoints out of web_server.py
26 routes and the similar-artists executor to api/artist_watchlist.py (-1,987
lines): the whole /api/watchlist surface plus its scan/timer machinery. The
handlers' shared helpers stay web_server-owned and are injected explicitly;
stable imports (validation, artist-source lookup, art-map cache, the registry's
spotify getter) come straight from their real homes instead. Verified by
url-map diff (1,152 routes, method+path identical) and an in-app smoke of
count + scan status.
